# backend/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Real Estate API",
    "description": "A robust and scalable RESTful API backend for a real estate marketplace. Endpoints are grouped by user role."
  },
  "tags": [
    {
      "name": "Guests",
      "description": "Operations available to anyone without an account."
    },
    {
      "name": "Buyers & Renters",
      "description": "Operations for authenticated users who are primarily browsing and saving properties."
    },
    {
      "name": "Sellers & Agents",
      "description": "Operations for authenticated users who manage property listings."
    },
    {
      "name": "Administrators",
      "description": "Operations for site administrators to manage users and content."
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Guests"],
        "summary": "Register a new user",
        "description": "Creates a new user account. The new user will have the 'buyer' role by default.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {"type": "string", "example": "jane_doe"},
                  "email": {"type": "string", "format": "email", "example": "jane@example.com"},
                  "password": {"type": "string", "example": "a_strong_password"}
                },
                "required": ["username", "email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": {"description": "User created successfully"},
          "400": {"description": "Bad Request or username/email already exists"}
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Guests"],
        "summary": "Log in a user",
        "description": "Authenticates a user and returns a JWT access token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {"type": "string", "example": "jane_doe"},
                  "password": {"type": "string", "example": "a_strong_password"}
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {"type": "string", "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."}
                  }
                }
              }
            }
          },
          "401": {"description": "Invalid credentials"}
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Buyers & Renters"],
        "summary": "Log out a user",
        "description": "Provides a confirmation for a client-side logout. The client should discard the JWT after receiving this response.",
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {"description": "Successfully logged out"},
          "401": {"description": "Unauthorized - Token is invalid or missing"}
        }
      }
    },
    "/properties": {
      "get": {
        "tags": ["Guests"],
        "summary": "Browse all properties",
        "description": "Retrieves a paginated list of all properties with optional filtering.",
        "parameters": [
          {"name": "page", "in": "query", "schema": {"type": "integer", "default": 1}, "example": 1},
          {"name": "per_page", "in": "query", "schema": {"type": "integer", "default": 10}, "example": 10},
          {"name": "location", "in": "query", "schema": {"type": "string"}, "example": "New York"},
          {"name": "property_type", "in": "query", "schema": {"type": "string"}, "example": "apartment"},
          {"name": "min_price", "in": "query", "schema": {"type": "number"}, "example": 100000},
          {"name": "max_price", "in": "query", "schema": {"type": "number"}, "example": 500000}
        ],
        "responses": {
          "200": {
            "description": "A paginated list of properties",
            "content": {"application/json": {"schema": {"type": "object"}}}
          }
        }
      },
      "post": {
        "tags": ["Sellers & Agents"],
        "summary": "Create a new property listing",
        "description": "Adds a new property to the database. Requires 'seller' or 'agent' role.",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "property_title": {"type": "string", "example": "Modern Downtown Apartment"},
                  "price": {"type": "number", "example": 550000},
                  "location": {"type": "string", "example": "New York, NY"},
                  "property_type": {"type": "string", "example": "apartment"},
                  "bedrooms": {"type": "integer", "example": 2},
                  "bathrooms": {"type": "integer", "example": 2},
                  "area_sqft": {"type": "integer", "example": 1200}
                },
                "required": ["property_title", "price", "location", "property_type", "bedrooms", "bathrooms", "area_sqft"]
              }
            }
          }
        },
        "responses": {
          "201": {"description": "Property created successfully"},
          "401": {"description": "Unauthorized - No token provided"},
          "403": {"description": "Forbidden - User does not have the required role"}
        }
      }
    },
    "/properties/{property_id}": {
      "get": {
        "tags": ["Guests"],
        "summary": "Get a single property by ID",
        "description": "Retrieves the details of a specific property. This call also increments the property's view count.",
        "parameters": [
          {"name": "property_id", "in": "path", "required": true, "schema": {"type": "string"}, "example": "507f1f77bcf86cd799439011"}
        ],
        "responses": {
          "200": {"description": "Property details"},
          "404": {"description": "Property not found"}
        }
      },
      "put": {
        "tags": ["Sellers & Agents"],
        "summary": "Update a property listing",
        "description": "Updates the details of a property. Only the original seller/agent or an admin can update it.",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {"name": "property_id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "price": {"type": "number", "example": 575000},
                  "property_title": {"type": "string", "example": "Updated Apartment Title"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Property updated successfully"},
          "401": {"description": "Unauthorized"},
          "403": {"description": "Forbidden - Not the owner"},
          "404": {"description": "Property not found"}
        }
      },
      "delete": {
        "tags": ["Sellers & Agents"],
        "summary": "Delete a property listing",
        "description": "Removes a property from the database. Only the original seller/agent or an admin can delete it.",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {"name": "property_id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Property deleted successfully"},
          "401": {"description": "Unauthorized"},
          "403": {"description": "Forbidden - Not the owner"},
          "404": {"description": "Property not found"}
        }
      }
    },
    "/properties/favorites": {
      "get": {
        "tags": ["Buyers & Renters"],
        "summary": "Get current user's favorite properties",
        "description": "Retrieves a list of all properties the current user has marked as a favorite.",
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {"description": "A list of favorite properties"}
        }
      }
    },
    "/properties/favorites/{property_id}": {
      "post": {
        "tags": ["Buyers & Renters"],
        "summary": "Add a property to favorites",
        "description": "Adds a property to the authenticated user's list of favorites.",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {"name": "property_id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Property added to favorites"},
          "400": {"description": "Invalid ID or property already favorited"}
        }
      },
      "delete": {
        "tags": ["Buyers & Renters"],
        "summary": "Remove a property from favorites",
        "description": "Removes a property from the authenticated user's list of favorites.",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {"name": "property_id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Property removed from favorites"},
          "400": {"description": "Invalid ID or property not in favorites"}
        }
      }
    },
    "/properties/{property_id}/contact": {
      "post": {
        "tags": ["Buyers & Renters"],
        "summary": "Send a message to the property's seller",
        "description": "Sends a message to the user who listed the property. The seller can then view it in their inbox.",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {"name": "property_id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {"type": "string", "example": "Is this property still available?"}
                },
                "required": ["message"]
              }
            }
          }
        },
        "responses": {
          "201": {"description": "Message sent successfully"},
          "400": {"description": "Bad Request or trying to contact yourself"}
        }
      }
    },
    "/auth/profile": {
      "get": {
        "tags": ["Buyers & Renters"],
        "summary": "Get current user's profile",
        "description": "Retrieves the profile information for the authenticated user.",
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {"description": "User profile data"},
          "401": {"description": "Unauthorized"}
        }
      }
    },
    "/auth/delete": {
      "delete": {
        "tags": ["Buyers & Renters"],
        "summary": "Delete current user's account",
        "description": "Permanently deletes the authenticated user's account and all their associated data (like properties).",
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {"description": "Account deleted successfully"},
          "401": {"description": "Unauthorized"}
        }
      }
    },
    "/admin/properties": {
      "get": {
        "tags": ["Administrators"],
        "summary": "Get all properties for admin review",
        "description": "Retrieves all properties in the system, with optional filters for admin management (e.g., reported properties).",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {"name": "status", "in": "query", "schema": {"type": "string", "example": "reported"}},
          {"name": "seller_id", "in": "query", "schema": {"type": "string", "example": "seller_jane"}}
        ],
        "responses": {
          "200": {"description": "A list of all properties"},
          "403": {"description": "Forbidden - Admin access required"}
        }
      }
    },
    "/admin/properties/{property_id}": {
      "delete": {
        "tags": ["Administrators"],
        "summary": "Delete any property",
        "description": "Allows an admin to delete any property from the system.",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {"name": "property_id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Property deleted by admin"},
          "403": {"description": "Forbidden - Admin access required"}
        }
      }
    },
    "/admin/users": {
      "get": {
        "tags": ["Administrators"],
        "summary": "Get all registered users",
        "description": "Retrieves a list of all users in the system for management purposes.",
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {"description": "A list of all users"},
          "403": {"description": "Forbidden - Admin access required"}
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter the token with the `Bearer ` prefix, e.g. 'Bearer eyJhbGciOi...'"
      }
    }
  }
}